requirements.txt
google-genai
huggingface_hub
Pillow
numpy
gTTS
edge-tts
requests
moviepy
wikipedia-api
beautifulsoup4
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
pydub
scikit-learn
imagehash
